+++ tstests/src/app.ts
@@ -15,10 +15,12 @@ log.info("App started on port : 3000");
 
 log.debug("UpdateOne(user=4)");
 
-log.debug("UpdateOne(user=4)", { id:4554645 });
+log.debug("UpdateOne(user=4)", { requestId:4554645 });
 
-log.debug("UpdateOne(user=4)", { id:4554645, userId: 45 });
+log.debug("UpdateOne(user=4)", { requestId:4554645, userId: 65.6 });
 
-log.debug("UpdateOne(user=4)", { id:4554645, userId: 45, tags: [ "update" ] });
+log.debug("UpdateOne(user=4)", { requestId:4554645, userId: 45 });
 
-log.debug("UpdateOne(user=4)", { id:4554645, userId: 45, tags:  [ "update", "user" ] });
+log.debug("UpdateOne(user=4)", { requestId:4554645, userId: 45, tags: [ "update" ] });
+
+log.debug("UpdateOne(user=4)", { requestId:4554645, userId: 45, tags:  [ "update", "user" ] });